Fixed .NET Core Hosting Bundle Install Script

+++ runbook/Install-NetCoreHostingBundle.ps1
@@ -22,35 +22,53 @@ process {
     $ErrorActionPreference = "Stop"
     $ProgressPreference = "SilentlyContinue"
 
-    Write-Verbose "Importing BitsTransfer Module"
-    Import-Module -Name "BitsTransfer" -Verbose:$false 4>&1 |
+    Write-Host "Importing BitsTransfer Module"
+    Import-Module -Name "BitsTransfer" |
     Out-Null
 
-    Write-Verbose "Downloading Installer"
+    Write-Host "Starting Bits Transfer"
     $destination = "C:\installers\$Msi"
     Write-Debug "`$destination = $destination"
-    Start-BitsTransfer -Source $Source -Destination $destination
 
-    Write-Verbose "Starting Installer"
+    try {
+        Start-BitsTransfer -Source $Source -Destination $destination -ErrorAction Stop
+    }
+    catch {
+        throw "Failed To Transfer Bits Bits Transfer From $Source ; $_"
+    }
+
+    if (-not (Test-Path -Path $destination)) {
+        throw "Failed To Find Installer At $destination"
+    }
+
+    Write-Host "Starting Installer"
     $parameters = @{
         FilePath         = "msiexec.exe"
         ArgumentList     = @(
-            "/i",
+            "/install",
             $destination,
             "/quiet",
             "/norestart"
         )
         NoNewWindow      = $true
+        PassThru         = $true
         Wait             = $true
         WorkingDirectory = $env:TEMP
     }
     Write-Debug "`$parameters = $($parameters | Out-String)"
-    Write-Warning "Downloaded ; Not Installed" -WarningAction Continue
-    # Start-Process @parameters
+    $process = Start-Process @parameters
 
+    if ($process.ExitCode -ne 0) {
+        throw "Failed To Install With Exit Code $($process.ExitCode)"
+    }
+
+    Write-Host "Asserting Cleanup"
     if ($Cleanup) {
-        Write-Verbose "Removing Installer"
+        Write-Host "Removing Installer"
         Remove-Item -Path $destination
+    }
+    else {
+        Write-Host "Skipping Cleanup"
     }
 }
 end {
